requirements.txt
matplotlib
numpy
opencv-python
Pillow
pyaml
scipy
tqdm
torch
torchaudio
torchvision
tensorboard
tiler
pyproj
pytorch_lightning
albumentations